+++ lib/packages.js
@@ -2,10 +2,11 @@ var fs = require('fs');
 
 module.exports = function(features) {
   var packages = {
-    'container':                  { trees: null,  requirements: [] },
+    'container':                  { trees: null,  requirements: ['ember-utils'] },
     'ember-environment':          { trees: null,  requirements: [], skipTests: true },
+    'ember-utils':                { trees: null,  requirements: [] },
     'ember-console':              { trees: null,  requirements: [], skipTests: true },
-    'ember-metal':                { trees: null,  requirements: ['ember-environment'], vendorRequirements: ['backburner'] },
+    'ember-metal':                { trees: null,  requirements: ['ember-environment', 'ember-utils'], vendorRequirements: ['backburner'] },
     'ember-debug':                { trees: null,  requirements: ['ember-metal'], testing: true },
     'ember-runtime':              { trees: null,  vendorRequirements: ['rsvp'], requirements: ['container', 'ember-environment', 'ember-console', 'ember-metal'] },
     'ember-views':                { trees: null,  requirements: ['ember-runtime'] },
