+++ packages/ember-metal/lib/events.js
@@ -7,7 +7,7 @@
 @submodule ember-metal
 */
 import { assert } from './debug';
-import { applyStr } from './utils';
+import { applyStr } from 'ember-utils';
 import { meta as metaFor, peekMeta } from './meta';
 import { deprecate } from './debug';
