+++ packages/ember-metal/lib/index.js
@@ -46,7 +46,7 @@ export {
   tryInvoke,
   uuid,
   wrap
-} from './utils';
+} from 'ember-utils';
 export {
   isTesting,
   setTesting
