+++ packages/ember-utils/lib/assign.js
