+++ packages/ember-utils/lib/empty-object.js
