+++ packages/ember-utils/lib/lookup-descriptor.js
@@ -1,0 +1,14 @@
+export default function lookupDescriptor(obj, keyName) {
+  let current = obj;
+  while (current) {
+    let descriptor = Object.getOwnPropertyDescriptor(current, keyName);
+
+    if (descriptor) {
+      return descriptor;
+    }
+
+    current = Object.getPrototypeOf(current);
+  }
+
+  return null;
+}
