+++ packages/ember-utils/tests/try_invoke_test.js
@@ -1,4 +1,4 @@
-import { tryInvoke } from '../../utils';
+import { tryInvoke } from '../index';
 
 let obj;
